refactor: clean department details
