+++ lib/config/ui_config.dart
@@ -25,8 +25,9 @@ abstract class CountdownConfig {
 }
 
 abstract class DepartmentsConfig {
-  static const defaultColorFirst = "#BFBEBE";
-  static const defaultColorSecond = "#999898";
+  // not used
+  // static const defaultColorFirst = "#BFBEBE";
+  // static const defaultColorSecond = "#999898";
 
   static const listSeparatorSize = 16.0;
   static const logoMicroPadding = 10.0;
