+++ lib/features/departments/department_detail_view/data/utils/field_of_studies_extension.dart
@@ -2,6 +2,12 @@ import "package:fast_immutable_collections/fast_immutable_collections.dart";
 
 import "../models/field_of_study.dart";
 
+extension CycleStudiesFormatterX on FieldOfStudy {
+  bool get isLongCycleStudies {
+    return semesterCount == longCycleStudiesSemesterCount;
+  }
+}
+
 extension WhereTypesX on IList<FieldOfStudy> {
   Iterable<FieldOfStudy> get whereFirstDegree {
     return where((item) => !item.is2ndDegree && !item.isLongCycleStudies);
