+++ lib/features/departments/department_detail_view/presentation/department_detail_view.dart
@@ -3,18 +3,18 @@ import "package:fast_immutable_collections/fast_immutable_collections.dart";
 import "package:flutter/material.dart";
 import "package:flutter_riverpod/flutter_riverpod.dart";
 
-import "../../../config/ui_config.dart";
-import "../../../services/translations_service/widgets/text_with_translation.dart";
-import "../../../theme/app_theme.dart";
-import "../../../utils/context_extensions.dart";
-import "../../../utils/determine_contact_icon.dart";
-import "../../../widgets/detail_views/contact_section.dart";
-import "../../../widgets/detail_views/detail_view_app_bar.dart";
-import "../../../widgets/horizontal_symmetric_safe_area.dart";
-import "../../../widgets/my_error_widget.dart";
-import "data/models/detailed_department.dart";
-import "data/repository/department_details_extension.dart";
-import "data/repository/department_details_repository.dart";
+import "../../../../config/ui_config.dart";
+import "../../../../services/translations_service/widgets/text_with_translation.dart";
+import "../../../../theme/app_theme.dart";
+import "../../../../utils/context_extensions.dart";
+import "../../../../utils/determine_contact_icon.dart";
+import "../../../../widgets/detail_views/contact_section.dart";
+import "../../../../widgets/detail_views/detail_view_app_bar.dart";
+import "../../../../widgets/horizontal_symmetric_safe_area.dart";
+import "../../../../widgets/my_error_widget.dart";
+import "../data/models/detailed_department.dart";
+import "../data/repository/department_details_repository.dart";
+import "../data/utils/department_details_extension.dart";
 import "widgets/fields_of_study_section.dart";
 import "widgets/science_clubs_section.dart";
 import "widgets/sliver_header_section.dart";
