+++ lib/features/departments/department_detail_view/presentation/widgets/view_loading.dart
@@ -1,10 +1,10 @@
 import "package:flutter/material.dart";
 
-import "../../../../config/ui_config.dart";
-import "../../../../widgets/loading_widgets/contact_section_loading.dart";
-import "../../../../widgets/loading_widgets/header_section_loading.dart";
-import "../../../../widgets/loading_widgets/shimmer_loading.dart";
-import "../../../home_view/widgets/loading_widgets/big_scrollable_section_loading.dart";
+import "../../../../../config/ui_config.dart";
+import "../../../../../widgets/loading_widgets/contact_section_loading.dart";
+import "../../../../../widgets/loading_widgets/header_section_loading.dart";
+import "../../../../../widgets/loading_widgets/shimmer_loading.dart";
+import "../../../../home_view/widgets/loading_widgets/big_scrollable_section_loading.dart";
 
 class DepartmentDetailViewLoading extends StatelessWidget {
   const DepartmentDetailViewLoading({super.key});
