--- lib/features/departments/department_detail_view/repository/getDepartmentDetails.graphql
@@ -1,24 +1,0 @@
-query GetDepartmentDetails($id: ID!) {
-    Departments_by_id(id: $id){
-        id
-        name
-        logo {
-            filename_disk
-        }
-        address
-        fieldsOfStudies {
-            name
-            is2ndDegree
-            isEnglish
-            isLongCycleStudies
-            hasWeekendModeOption
-            url
-        }
-        links {
-            name
-            link
-        }
-        gradient_start,
-        gradient_end
-    }
-}
